refactor(fe): verifyUserState 함수 및 유저인증 디스페치 리팩토링
